creating FragmentViewPanel
