+++ src/GUI/src/gui/GUIFragmentInspector.java
@@ -3,14 +3,12 @@ package gui;
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Dimension;
-import java.awt.Graphics;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.beans.PropertyChangeEvent;
+import java.beans.PropertyChangeListener;
 import java.io.File;
 import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import javax.swing.BoxLayout;
@@ -21,21 +19,13 @@ import javax.swing.JPanel;
 import javax.swing.JSeparator;
 import javax.swing.JSpinner;
 import javax.swing.JSpinner.DefaultEditor;
-import javax.swing.JTable;
 import javax.swing.SpinnerNumberModel;
 import javax.swing.SwingConstants;
 import javax.swing.UIManager;
 import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
-import javax.swing.event.TableModelEvent;
-import javax.swing.event.TableModelListener;
-import javax.swing.table.DefaultTableModel;
 import javax.vecmath.Point3d;
 
-import org.jmol.adapter.smarter.SmarterJmolAdapter;
-import org.jmol.api.JmolViewer;
-import org.jmol.viewer.Viewer;
-import org.openscience.cdk.AtomContainer;
 import org.openscience.cdk.interfaces.IAtom;
 import org.openscience.cdk.interfaces.IAtomContainer;
 import org.openscience.cdk.interfaces.IBond;
@@ -45,7 +35,6 @@ import denoptim.exception.DENOPTIMException;
 import denoptim.io.DenoptimIO;
 import denoptim.molecule.DENOPTIMAttachmentPoint;
 import denoptim.molecule.DENOPTIMFragment;
-import denoptim.utils.DENOPTIMMathUtils;
 import denoptim.utils.FragmentUtils;
 
 
@@ -81,27 +70,16 @@ public class GUIFragmentInspector extends GUICardPanel
 	private DENOPTIMFragment fragment;
 	
 	/**
-	 * Temporary list of attachment points of the current fragment
-	 */
-	private Map<Integer,DENOPTIMAttachmentPoint> mapAPs = null;
-	
-	/**
 	 * The index of the currently loaded fragment [0–(n-1)}
 	 */
 	private int currFrgIdx = 0;
-	
-	/**
-	 * Flag signaling that data about APs has been changed in the GUI
-	 */
-	private boolean alteredAPData = false;
 	
 	/**
 	 * Flag signaling that loaded data has changes since last save
 	 */
 	private boolean unsavedChanges = false;
 	
-	private JPanel centralPanel;
-	private JmolPanel jmolPanel;
+	private FragmentViewPanel fragmentViewer;
 	private JPanel fragCtrlPane;
 	private JPanel fragNavigPanel;
 	private JPanel fragNavigPanel2;
@@ -117,9 +95,6 @@ public class GUIFragmentInspector extends GUICardPanel
 	private final FragSpinnerChangeEvent fragSpinnerListener = 
 			new FragSpinnerChangeEvent();
 	
-	private DefaultTableModel apTabModel;
-	private JTable apTable;
-	
 	private JPanel pnlOpenMol;
 	private JButton btnOpenMol;
 	
@@ -132,15 +107,9 @@ public class GUIFragmentInspector extends GUICardPanel
 	private JPanel pnlDelSel;
 	private JButton btnDelSel;
 	
-	private JPanel pnlPrintMol; //for testing
-	private JButton btnPrintMol;//for testing
-	
 	private JPanel pnlSaveEdits;
 	private JButton btnSaveEdits;
-	
-	private final String NL = System.getProperty("line.separator");
-	
-	String tmpSDFFile = "/tmp/Denoptim_GUIFragInspector_loadedMol.sdf";
+
 	
 //-----------------------------------------------------------------------------
 	
@@ -159,49 +128,26 @@ public class GUIFragmentInspector extends GUICardPanel
 	/**
 	 * Initialize the panel and add buttons.
 	 */
-	@SuppressWarnings("serial")
 	private void initialize() {
 		
 		// BorderLayout is needed to allow dynamic resizing!
 		this.setLayout(new BorderLayout()); 
 		
 		// This card structure includes center, east and south panels:
-		// - (Center) Jmol stuff. Includes
-		//            - molecular viewer
-		//            - list (table) of attachment points
+		// - (Center) molecular viewer and APs
 		// - (East) Fragment controls
 		// - (South) general controls (load, save, close)
 		
-		// The Jmol stuff goes all in here
-        centralPanel = new JPanel(new BorderLayout());
-        this.add(centralPanel,BorderLayout.CENTER);
-        
-        // Jmol viewer panel
-        jmolPanel = new JmolPanel();
-        jmolPanel.setPreferredSize(new Dimension(400, 400));
-        centralPanel.add(jmolPanel,BorderLayout.CENTER);
-        
-		// List of attachment points
-		apTabModel = new DefaultTableModel(){
+		// The viewer with Jmol and APtable
+		fragmentViewer = new FragmentViewPanel(true);
+		fragmentViewer.addPropertyChangeListener("APDATA", 
+				new PropertyChangeListener() {
 			@Override
-		    public boolean isCellEditable(int row, int column) {
-				if (column == 0 || row == 0)
-				{
-					return false;
-				}
-				else
-			    {
-					return true;
-			    }
-		    }
-		};
-		apTabModel.setColumnCount(2);
-		apTable = new JTable(apTabModel);
-		apTable.putClientProperty("terminateEditOnFocusLost", true);
-		apTable.getColumnModel().getColumn(0).setMaxWidth(75);
-		apTable.setGridColor(Color.BLACK);
-		apTabModel.addTableModelListener(new PausableTableModelListener());
-		centralPanel.add(apTable,BorderLayout.SOUTH);
+			public void propertyChange(PropertyChangeEvent evt) {
+				protectEditedSystem();				
+			}
+		});
+        this.add(fragmentViewer,BorderLayout.CENTER);
 		
 		// General panel on the right: it containing all controls
         fragCtrlPane = new JPanel();
@@ -282,7 +228,7 @@ public class GUIFragmentInspector extends GUICardPanel
 					removeCurrentFragment();
 					if (fragmentLibrary.size()==0)
 					{
-						jmolPanel.viewer.evalString("zap");
+						fragmentViewer.clearMolecularViewer();
 					}
 				} catch (DENOPTIMException e1) {
 					System.out.println("Esception while removing the current fragment:");
@@ -342,7 +288,8 @@ public class GUIFragmentInspector extends GUICardPanel
 			    + "<br><b>WARNING:</b> this action cannot be undone!<html>");
 		btnAtmToAP.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				ArrayList<IAtom> selectedAtms = getAtomsSelectedFromJMol();
+				ArrayList<IAtom> selectedAtms = 
+						fragmentViewer.getAtomsSelectedFromJMol();
 				
 				if (selectedAtms.size() == 0)
 				{
@@ -356,9 +303,6 @@ public class GUIFragmentInspector extends GUICardPanel
 				}
 				else
 				{
-					//TODO: deal with changes to the AP table first
-					
-					
 					//TODO: ask about hapticity:
 					// if multihapto, then use all the selected atoms for 1 AP
 					// and ask to select another set for other end of bond to 
@@ -388,18 +332,10 @@ public class GUIFragmentInspector extends GUICardPanel
 					// Use the APs stored in the atoms
 					fragment.updateAPs();
 					
-					// Load the "fragment" obj to the viewer
-					loadCurrentAsPlainStructure();
-					
-					// Update APs using info in atoms (overwrites map and table)
-					updateAPsMapAndTable();
-			        
-			        // Display the APs as arrows
-			        updateAPsInJmolViewer();
+					fragmentViewer.loadFragImentToViewer(fragment);
 					
 			        // Protect the temporary "fragment" obj
 			        unsavedChanges = true;
-			        alteredAPData = true;
 			        protectEditedSystem();
 				}
 			}
@@ -414,7 +350,8 @@ public class GUIFragmentInspector extends GUICardPanel
 				+ "undone!");
 		btnDelSel.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				ArrayList<IAtom> selectedAtms = getAtomsSelectedFromJMol();
+				ArrayList<IAtom> selectedAtms = 
+						fragmentViewer.getAtomsSelectedFromJMol();
 				
 				if (selectedAtms.size() == 0)
 				{
@@ -427,61 +364,21 @@ public class GUIFragmentInspector extends GUICardPanel
 					return;
 				}
 				else
-				{
-					//TODO: deal with changes to the AP table first
-					
+				{					
 					removeAtoms(selectedAtms);
 					
-					// Use the APs stored in the atoms
 					fragment.updateAPs();
 					
-					// Load the "fragment" obj to the viewer
-					loadCurrentAsPlainStructure();
-					
-					// Update APs using info in atoms (overwrites map and table)
-					updateAPsMapAndTable();
-			        
-			        // Display the APs as arrows
-			        updateAPsInJmolViewer();
+					fragmentViewer.loadFragImentToViewer(fragment);
 					
 			        // Protect the temporary "fragment" obj
 			        unsavedChanges = true;
-			        alteredAPData = true;
 			        protectEditedSystem();
 				}
 			}
 		});
 		pnlDelSel.add(btnDelSel);
 		fragCtrlPane.add(pnlDelSel);
-		
-		//TODO comment out button (only for testing)
-		/*
-        pnlPrintMol = new JPanel();
-        btnPrintMol = new JButton("TEST STUFF");
-        btnPrintMol.addActionListener(new ActionListener() {
-            public void actionPerformed(ActionEvent e) {         	
-            	System.out.println("#Bonds: "+fragment.getBondCount());
-            	System.out.println("#Atoms: "+fragment.getAtomCount());
-            	for (IAtom atm : fragment.atoms())
-            	{
-                	System.out.println("    "+atm.getSymbol()+fragment.getAtomNumber(atm));
-            	}
-            	System.out.println("#APs:   "+fragment.getAPCount());
-            	for (DENOPTIMAttachmentPoint ap : fragment.getCurrentAPs())
-            	{
-            		System.out.println("    "+ap);
-            	}
-            	System.out.println("Library");
-            	for (IAtomContainer mol : fragmentLibrary)
-            	{
-            		System.out.println("Frag #atoms: "+mol.getAtomCount());
-            	}
-            	
-            }
-        });
-        pnlPrintMol.add(btnPrintMol);
-        fragCtrlPane.add(pnlPrintMol);
-		*/
 		
 		fragCtrlPane.add(new JSeparator());
 		
@@ -628,48 +525,15 @@ public class GUIFragmentInspector extends GUICardPanel
 		clearCurrentSystem();
 		
 		// Load the structure using CACTUS service via Jmol
-		jmolPanel.viewer.evalString("load $"+smiles);
-		waitForJmolViewer(1500);
-
-		// NB: this is a workaround to the lack fo try/catch mechanism when
-		// executing Jmol commands.
-		// We want to catch errors that prevent loading the structure
-		// For example, offline mode and invalid SMILES
-		String data = jmolPanel.viewer.getData("*", "txt");				
-		if (data == null || data.equals(""))
-		{
-			System.out.println("DATA: "+data);
-			JOptionPane.showMessageDialog(null,
-	                "<html>Could not find a valid structure.<br>"
-	                + "Possible reasons are:"
-	                + "<ul>"
-	                + "<li>unreachable remote service (we are offline)</li>"
-	                + "<li>no available structure for SMILES string<br>'" 
-	                + smiles 
-	                + "'</li></ul><br>"
-	                + "Please create such structure and import it as SDF file."
-	                + "</html>",
-	                "Error",
-	                JOptionPane.PLAIN_MESSAGE,
-	                UIManager.getIcon("OptionPane.errorIcon"));
-			return;
-		}
-		
-		// Now we should have a structure loaded in the viewer, 
-		// so we take that one and put it in the IAtomContainer representation
-		try	{
-		    fragment = new DENOPTIMFragment(getStructureFromJmolViewer());
+		try {
+			fragmentViewer.loadSMILESFromRemote(smiles);
 		} catch (Exception e) {
-			e.printStackTrace();
-			JOptionPane.showMessageDialog(null,
-	                "<html>Could not understand Jmol system.</html>",
-	                "Error",
-	                JOptionPane.PLAIN_MESSAGE,
-	                UIManager.getIcon("OptionPane.errorIcon"));
 			return;
 		}
 		
-		// the system is not a fragment but, this is done for consistency:
+		fragment = fragmentViewer.getLoadedStructure();
+		
+		// The system is not a fragment but, this is done for consistency:
 		// when we have a molecule loaded the list is not empty:
 		// The currently viewed fragment (if any) is always part of the library
 	    fragmentLibrary.add(fragment);
@@ -677,104 +541,7 @@ public class GUIFragmentInspector extends GUICardPanel
 		
 		// finalize GUI status
 		updateFragListSpinner();
-		setJmolViewer();
 		unsavedChanges = true;
-	}
-	
-//-----------------------------------------------------------------------------
-	
-	private IAtomContainer getStructureFromJmolViewer() throws DENOPTIMException
-	{
-		IAtomContainer mol = new AtomContainer();
-		
-		String strData = jmolPanel.viewer.getData("*", "txt");
-		strData = strData.replaceAll("[0-9] \\s+ 999 V2000", 
-				"0  0  0  0  0999 V2000");
-		String[] lines = strData.split("\\n");
-		if (lines.length < 5)
-		{
-			clearCurrentSystem();
-			throw new DENOPTIMException("Unexpected format in Jmol molecular "
-					+ "data: " + strData);
-		}
-		if (!lines[3].matches(".*999 V2000.*"))
-		{
-			clearCurrentSystem();
-			throw new DENOPTIMException("Unexpected format in Jmol molecular "
-					+ "data: " + strData);
-		}
-		String[] counters = lines[3].trim().split("\\s+");
-		int nAtms = Integer.parseInt(counters[0]);
-		int nBonds = Integer.parseInt(counters[1]);
-		
-		
-		StringBuilder sb = new StringBuilder();
-		sb.append("Structure in JmolViewer").append(NL);
-		sb.append("Jmol").append(NL).append(NL);
-		sb.append(lines[3]).append(NL);
-		for (int i=0; i<nAtms; i++)
-		{
-			sb.append(lines[3+i+1]).append("  0  0  0  0  0  0").append(NL);
-		}
-		for (int i=0; i<nBonds; i++)
-		{
-			sb.append(lines[3+nAtms+i+1]).append("  0").append(NL);
-		}
-		sb.append("M  END").append(NL).append("$$$$");
-		
-		DenoptimIO.writeData(tmpSDFFile, sb.toString(), false);
-		mol = DenoptimIO.readSingleSDFFile(tmpSDFFile);
-		return mol;
-	}
-	
-//-----------------------------------------------------------------------------
-	
-	private void waitForJmolViewer(int milliSecFirst)
-	{
-		// We wait
-		try {
-			Thread.sleep(milliSecFirst);
-		} catch (InterruptedException e) {
-			e.printStackTrace();
-		}
-		
-		// Waiting cycles
-		Date date = new Date();
-		long startTime = date.getTime();
-		long wallTime = startTime + 5000; // 5 seconds
-		while (jmolPanel.viewer.isScriptExecuting())
-		{
-				
-			Date newDate = new Date();
-			long now = newDate.getTime();
-			System.out.println("Waiting "+now);
-			if (now > wallTime)
-			{
-				String[] options = new String[]{"Yes","No"};
-				int res = JOptionPane.showOptionDialog(null,
-		                "<html>Slow response from Jmol.<br>Keep waiting?</html>",
-		                "Should we wait for another 5 seconds?",
-		                JOptionPane.DEFAULT_OPTION,
-		                JOptionPane.QUESTION_MESSAGE,
-		                UIManager.getIcon("OptionPane.warningIcon"),
-		                options,
-		                options[1]);
-				if (res == 1)
-				{
-					break;
-				}
-				else
-				{
-					newDate = new Date();
-					wallTime = newDate.getTime() + 5000; // 5 seconds
-				}
-			}
-			try {
-				Thread.sleep(500);
-			} catch (InterruptedException e) {
-				e.printStackTrace();
-			}
-		}
 	}
 	
 //-----------------------------------------------------------------------------
@@ -840,29 +607,7 @@ public class GUIFragmentInspector extends GUICardPanel
 	 */
 	private void loadCurrentAsPlainStructure()
 	{
-		if (fragment == null)
-		{
-			JOptionPane.showMessageDialog(null,
-	                "<html>No structure loaded.<br>This is most likely a bug!"
-	                + "Please report it to the development team.</html>",
-	                "Error",
-	                JOptionPane.PLAIN_MESSAGE,
-	                UIManager.getIcon("OptionPane.errorIcon"));
-			return;			
-		}
-		
-		// NB: we use the nasty trick of a tmp file to by-pass the 
-		// fragile/discontinued CDK-to-Jmol support.
-		
-		try {
-			DenoptimIO.writeMolecule(tmpSDFFile, fragment, false);
-		} catch (DENOPTIMException e) {
-			// TODO change and look for other tmp file locations
-			e.printStackTrace();
-			System.out.println("Error writing TMP file");
-		}
-		jmolPanel.viewer.openFile(tmpSDFFile);
-		setJmolViewer();
+		fragmentViewer.loadPlainStructure(fragment);
 	}
 
 //-----------------------------------------------------------------------------
@@ -887,62 +632,9 @@ public class GUIFragmentInspector extends GUICardPanel
 		}
 		
 		clearCurrentSystem();
-			
-		// Load the molecular data into Jmol
-		// NB: we use the nasty trick of a tmp file to by-pass the 
-		// fragile/discontinued CDK-to-Jmol support.
+
 		fragment = fragmentLibrary.get(currFrgIdx);
-		loadCurrentAsPlainStructure();
-		
-		// Update APs using info in atoms (overwrites map and table)
-		updateAPsMapAndTable();
-        
-        // Display the APs as arrows
-        updateAPsInJmolViewer();
-	}
-	
-//-----------------------------------------------------------------------------
-	
-	/**
-	 * Uses the AP of the {@link DENOPTIMFragment} to create a new map and 
-	 * table of APs.
-	 */
-	private void updateAPsMapAndTable()
-	{
-		clearAPTable();
-		mapAPs = new HashMap<Integer,DENOPTIMAttachmentPoint>();
-		
-		ArrayList<DENOPTIMAttachmentPoint> lstAPs = fragment.getCurrentAPs();		
-        if (lstAPs.size() == 0)
-        {
-        	// WARNING: no dialog here because it fires event changes in JSpinner
-			return;
-        }
-        
-        activateTabEditsListener(false);
-        apTabModel.addRow(new Object[]{"<html><b>AP#<b></html>",
-		"<html><b>APClass<b></html>"});
-        
-        int arrId = 0;
-	    for (DENOPTIMAttachmentPoint ap : lstAPs)
-	    {
-	    	arrId++;
-	    	apTabModel.addRow(new Object[]{arrId, ap.getAPClass()});
-	    	mapAPs.put(arrId,ap);
-	    }
-	    activateTabEditsListener(true);
-	}
-	
-//-----------------------------------------------------------------------------
-	
-	private void clearAPTable()
-	{
-		activateTabEditsListener(false);
-        for (int i=(apTabModel.getRowCount()-1); i>-1; i--) 
-        {
-        	apTabModel.removeRow(i);
-        }
-        activateTabEditsListener(true);
+		fragmentViewer.loadFragImentToViewer(fragment);
 	}
 	
 //-----------------------------------------------------------------------------
@@ -958,48 +650,10 @@ public class GUIFragmentInspector extends GUICardPanel
 		// submitting "zap" only in that occasion.
 		
 		// Remove tmp storage of APs
-		mapAPs = null;
+		fragmentViewer.mapAPs = null;
 		
 		// Remove table of APs
-		clearAPTable();
-	}
-
-//-----------------------------------------------------------------------------
-	
-	private void updateAPsInJmolViewer()
-	{   
-		if (mapAPs == null || mapAPs.isEmpty())
-		{
-			return;
-		}
-		
-		StringBuilder sb = new StringBuilder();
-        for (int arrId : mapAPs.keySet())
-        {
-        	DENOPTIMAttachmentPoint ap = mapAPs.get(arrId);
-        	int srcAtmId = ap.getAtomPositionNumber();
-        	Point3d srcAtmPlace = FragmentUtils.getPoint3d(
-        			fragment.getAtom(srcAtmId));
-        	double[] startArrow = new double[]{
-        			srcAtmPlace.x,
-        			srcAtmPlace.y,
-        			srcAtmPlace.z};
-        	double[] endArrow = ap.getDirectionVector();
-        	
-        	double[] offSet = DENOPTIMMathUtils.scale(DENOPTIMMathUtils.subtract(endArrow,startArrow), 0.2);
-        	double[] positionLabel = DENOPTIMMathUtils.add(endArrow,offSet); 
-        	sb.append("draw arrow").append(arrId).append(" arrow ");
-        	sb.append(getJmolPositionStr(startArrow));
-        	sb.append(getJmolPositionStr(endArrow));
-        	sb.append(" width 0.1");
-        	sb.append(NL);
-        	sb.append("set echo apLab").append(arrId);
-        	sb.append(getJmolPositionStr(positionLabel));
-        	sb.append("; echo ").append(arrId);
-        	sb.append("; color echo yellow");
-        	sb.append(NL);
-        }
-        jmolPanel.viewer.evalString(sb.toString());
+		fragmentViewer.clearAPTable();
 	}
 
 //-----------------------------------------------------------------------------
@@ -1010,84 +664,6 @@ public class GUIFragmentInspector extends GUICardPanel
 				fragmentLibrary.size(), 1));
 		totalFragsLabel.setText(Integer.toString(fragmentLibrary.size()));
 	}
-	
-//-----------------------------------------------------------------------------
-	
-	private String getJmolPositionStr(double[] position)
-	{
-		StringBuilder sb = new StringBuilder();
-    	sb.append(" {");
-    	sb.append(position[0]).append(" ");
-    	sb.append(position[1]).append(" ");
-    	sb.append(position[2]).append("} ");
-		return sb.toString();
-	}
-	
-//-----------------------------------------------------------------------------
-	
-	private void setJmolViewer()
-	{
-		//TODO del
-		System.out.println("In setJmolViewer");
-		StringBuilder sb = new StringBuilder();		
-		sb.append("select none").append(NL); 
-		sb.append("SelectionHalos ON").append(NL);
-		sb.append("set picking ATOMS").append(NL);
-		//sb.append("zoom 70").append(NL);
-		
-		jmolPanel.viewer.evalString(sb.toString());
-
-	}
-	
-//-----------------------------------------------------------------------------
-	
-	private class JmolPanel extends JPanel 
-	{
-
-        /**
-		 * Version UID
-		 */
-		private static final long serialVersionUID = 1699908697703788097L;
-
-		JmolViewer viewer;
-
-        private final Dimension hostPanelSize = new Dimension();
-
-        public JmolPanel() {
-            viewer = JmolViewer.allocateViewer(this, new SmarterJmolAdapter(), 
-            null, null, null, null, null); //NB: can add listener here
-        }
-        
-        //---------------------------------------------------------------------
-
-		@Override
-        public void paint(Graphics g) {
-            getSize(hostPanelSize);
-            viewer.renderScreenImage(g, hostPanelSize.width, hostPanelSize.height);
-        }
-    }
-
-//-----------------------------------------------------------------------------
-    
-    private ArrayList<IAtom> getAtomsSelectedFromJMol()
-    {
-    	ArrayList<IAtom> selectedAtms = new ArrayList<IAtom>();
-    	
-    	if (fragment == null || fragment.getAtomCount()==0)
-    	{
-    		return selectedAtms;
-    	}
-    	
-		for (int i =0; i< fragment.getAtomCount(); i++)
-		{
-			if (((Viewer) jmolPanel.viewer).slm.isSelected(i))
-			{
-				selectedAtms.add(fragment.getAtom(i));
-			}
-		}
-    	
-		return selectedAtms;
-    }
     
 //-----------------------------------------------------------------------------
     
@@ -1145,47 +721,6 @@ public class GUIFragmentInspector extends GUICardPanel
     	return true;
     }
     
-//-----------------------------------------------------------------------------
-    
-    /**
-     * Create a multihapto AP from centroids of set of atoms.
-     * @param srcAtms the set to be used to calculate centroid of the start of
-     * the formal bond to break, or the end atom if only one is present
-     * @param endAtm the set to be used to calculate centroid of the end of the
-     * formal bond to break, or the end atom if only one is present
-     * 
-     */
-    private void createAP(IAtom[] srcAtms, IAtom[] endAtms)
-    {
-    	//TODO
-    }
-    
-//-----------------------------------------------------------------------------
-    
-    /**
-     * Create a simple AP considering the bond from one atom to another.
-     * @param srcAtm the atom on the fragment, and that will hold the AP
-     * @param trgAtm the atom representing the next fragment, which does not 
-     * exist. This atom will be removed from the system.
-     */
-    private void createAP(IAtom srcAtm, IAtom trgAtm)
-    {  		
-    	//Ask about class
-    	//TODO
-    	
-    	//Make AP (use half-bond position for AP end)
-    	//TODO
-    	
-    	// Remove atom
-    	//TODO
-    	
-    	// Update APs
-    	//TODO
-    	
-    	//Update viewer
-    	//TODO
-    }
-    
 //----------------------------------------------------------------------------
 
     private void removeAtoms(ArrayList<IAtom> atmsToDels)
@@ -1212,31 +747,6 @@ public class GUIFragmentInspector extends GUICardPanel
     	
     	
     }
-	
-//-----------------------------------------------------------------------------
-	
-	private class PausableTableModelListener implements TableModelListener
-	{	
-		private boolean isActive = false;
-		
-		public PausableTableModelListener() {};
-
-		@Override
-		public void tableChanged(TableModelEvent e) 
-		{
-            if (isActive && !alteredAPData 
-            		&& e.getType() == TableModelEvent.UPDATE)
-            {
-                alteredAPData = true;
-                protectEditedSystem();
-            }
-		}
-        
-		public void setActive(boolean var)
-		{
-			isActive = var;
-		}
-	}
 	
 //-----------------------------------------------------------------------------
 	
@@ -1284,6 +794,8 @@ public class GUIFragmentInspector extends GUICardPanel
 		btnOpenFrags.setEnabled(true);
 		btnOpenSMILES.setEnabled(true); 
 		btnOpenMol.setEnabled(true);
+		btnDelSel.setEnabled(true);
+		btnAtmToAP.setEnabled(true);
 		
 		fragNavigSpinner.setModel(new SpinnerNumberModel(currFrgIdx+1, 1, 
 				fragmentLibrary.size(), 1));
@@ -1291,6 +803,7 @@ public class GUIFragmentInspector extends GUICardPanel
 			.getTextField().setEditable(true); 
 		((DefaultEditor) fragNavigSpinner.getEditor())
 			.getTextField().setForeground(Color.BLACK);
+		fragmentViewer.deprotectEdits();
 		
 		fragSpinnerListener.setEnabled(true);
 	}
@@ -1304,6 +817,8 @@ public class GUIFragmentInspector extends GUICardPanel
 		btnOpenFrags.setEnabled(false);
 		btnOpenSMILES.setEnabled(false); 
 		btnOpenMol.setEnabled(false);
+		btnDelSel.setEnabled(false);
+		btnAtmToAP.setEnabled(false);
 		
 		fragNavigSpinner.setModel(new SpinnerNumberModel(currFrgIdx+1, 
 				currFrgIdx+1, currFrgIdx+1, 1));
@@ -1319,24 +834,14 @@ public class GUIFragmentInspector extends GUICardPanel
 
     private void activateTabEditsListener(boolean var)
     {
-		try
-		{
-			PausableTableModelListener l = (PausableTableModelListener) 
-					apTabModel.getTableModelListeners()[0];
-    	    l.setActive(var);
-		} catch (Throwable t) {
-			//t.printStackTrace();
-			System.out.println("Bad attempt to contro llistener: " 
-					+ t.getMessage());
-			System.out.println(t.getCause());
-		}
+		fragmentViewer.activateTabEditsListener(var);
     }
     
 //-----------------------------------------------------------------------------
     
     private void removeCurrentFragment() throws DENOPTIMException
     {
-    	if (alteredAPData)
+    	if (fragmentViewer.hasUnsavedAPEdits())
     	{
 			String[] options = new String[]{"Yes","No"};
 			int res = JOptionPane.showOptionDialog(null,
@@ -1376,7 +881,6 @@ public class GUIFragmentInspector extends GUICardPanel
 	    	updateFragListSpinner();
 	    	
 	  		// Release constraints
-	        alteredAPData = false;
 	        deprotectEditedSystem();
 	        
     	}
@@ -1386,87 +890,61 @@ public class GUIFragmentInspector extends GUICardPanel
 
   	private void saveUnsavedChanges() 
   	{
-  		if (!alteredAPData)
+  		if (fragmentViewer.hasUnsavedAPEdits())
   		{
-  			return;
-  		}
-	    
-  		// Retrieve chemical object from Jmol
-		try	{
-		    fragment = new DENOPTIMFragment(getStructureFromJmolViewer());
-		} catch (Exception e) {
-			e.printStackTrace();
-			JOptionPane.showMessageDialog(null,
-	                "<html>Could not understand Jmol system.</html>",
-	                "Error",
-	                JOptionPane.ERROR_MESSAGE,
-	                UIManager.getIcon("OptionPane.errorIcon"));
-			return;
-		}
-  		
-  		// Import changes from AP table into molecular representation
-        for (int i=1; i<apTabModel.getRowCount(); i++) 
-        {
-        	int apId = ((Integer) apTabModel.getValueAt(i, 0)).intValue();
-        	String currApClass = apTabModel.getValueAt(i, 1).toString();
-        	
-        	// Make sure the new class has a proper syntax
-        	currApClass = ensureGoodAPClassString(currApClass);
-        	
-        	if (mapAPs.containsKey(apId))
-        	{
-        		String origApClass = mapAPs.get(apId).getAPClass();
-        		if (!origApClass.equals(currApClass))
-        		{
-        			try {
-						mapAPs.get(apId).setAPClass(currApClass);
-					} catch (DENOPTIMException e) {
-						// We made sure the class is valid, so this
-						// should never happen, though one never knows
-						e.printStackTrace();
-						JOptionPane.showMessageDialog(null,
-		    	                "<html>Could not save due to errors setting a "
-		    	                + "new APClass.<br>Please report this to the "
-		    	                + "DENOPTIM team.</html>",
-		    	                "Error",
-		    	                JOptionPane.PLAIN_MESSAGE,
-		    	                UIManager.getIcon("OptionPane.errorIcon"));
-		    			return;	
-					}
-        		}
-        	}
-        	else
-        	{
-    			JOptionPane.showMessageDialog(null,
-    	                "<html>Could not save due to mistmatch between AP "
-    	                + "table and map.<br>Please report this to the "
-    	                + "DENOPTIM team.</html>",
-    	                "Error",
-    	                JOptionPane.PLAIN_MESSAGE,
-    	                UIManager.getIcon("OptionPane.errorIcon"));
-    			return;	
-        	}
-        	
-        	// Add AP to molecular representation
-        	DENOPTIMAttachmentPoint ap = mapAPs.get(apId);
-        	try {
-				fragment.addAP(ap.getAtomPositionNumber(), currApClass,
-						ap.getDirectionVector());
-			} catch (DENOPTIMException e) {
-				e.printStackTrace();
-    			JOptionPane.showMessageDialog(null,
-    	                "<html>Could not add attachment points.<br>"
-    	                + "Please report this to the "
-    	                + "DENOPTIM team.</html>",
-    	                "Error",
-    	                JOptionPane.PLAIN_MESSAGE,
-    	                UIManager.getIcon("OptionPane.errorIcon"));
-    			return;	
-			}
-        }
-  		
-  		// Overwrite fragment in library
-  		fragmentLibrary.set(currFrgIdx, fragment);
+	  		// Retrieve chemical object from the viewer
+	  		fragment = fragmentViewer.getLoadedStructure();
+	  		
+	  		// Import changes from AP table into molecular representation
+	        for (int i=1; i<fragmentViewer.apTabModel.getRowCount(); i++) 
+	        {	        	
+	        	int apId = ((Integer) fragmentViewer.apTabModel.getValueAt(i, 0))
+	        			.intValue();
+	        	String currApClass = fragmentViewer.apTabModel.getValueAt(i, 1)
+	        			.toString();
+	        	
+	        	// Make sure the new class has a proper syntax
+	        	currApClass = ensureGoodAPClassString(currApClass);
+	        	
+	        	if (fragmentViewer.mapAPs.containsKey(apId))
+	        	{
+	        		String origApClass = 
+	        				fragmentViewer.mapAPs.get(apId).getAPClass();
+	        		if (!origApClass.equals(currApClass))
+	        		{
+	        			try {
+	        				fragmentViewer.mapAPs.get(apId).setAPClass(currApClass);
+						} catch (DENOPTIMException e) {
+							// We made sure the class is valid, so this
+							// should never happen, though one never knows
+							e.printStackTrace();
+							JOptionPane.showMessageDialog(null,
+			    	                "<html>Could not save due to errors setting a "
+			    	                + "new APClass.<br>Please report this to the "
+			    	                + "DENOPTIM team.</html>",
+			    	                "Error",
+			    	                JOptionPane.PLAIN_MESSAGE,
+			    	                UIManager.getIcon("OptionPane.errorIcon"));
+			    			return;	
+						}
+	        		}
+	        	}
+	        	else
+	        	{
+	    			JOptionPane.showMessageDialog(null,
+	    	                "<html>Could not save due to mistmatch between AP "
+	    	                + "table and map.<br>Please report this to the "
+	    	                + "DENOPTIM team.</html>",
+	    	                "Error",
+	    	                JOptionPane.PLAIN_MESSAGE,
+	    	                UIManager.getIcon("OptionPane.errorIcon"));
+	    			return;	
+	        	}
+	        }
+	  		
+	  		// Overwrite fragment in library
+	  		fragmentLibrary.set(currFrgIdx, fragment);
+	  	}
         
         // Reload fragment from library to refresh table and viewer
     	activateTabEditsListener(false);
@@ -1474,7 +952,6 @@ public class GUIFragmentInspector extends GUICardPanel
   		
   		// Release constraints
     	activateTabEditsListener(true);
-        alteredAPData = false;
         deprotectEditedSystem();
   	}
